--- a/kolibri/plugins/coach/api.py
+++ b/kolibri/plugins/coach/api.py
@@ -1,9 +1,8 @@
 import datetime
-from dateutil.parser import parse
 
+from dateutil.parser import parse
 from django.db.models import Min, Q
 from django.utils import timezone
-
 from kolibri.auth.constants import role_kinds
 from kolibri.auth.models import Collection, FacilityUser
 from kolibri.content.models import ContentNode
@@ -57,7 +56,7 @@ class ContentReportViewSet(viewsets.ModelViewSet):
 
     def get_queryset(self):
         content_node_id = self.kwargs['content_node_id']
-        return ContentNode.objects.filter(parent=content_node_id)
+        return ContentNode.objects.filter(parent=content_node_id).order_by('lft')
 
 
 class ContentSummaryViewSet(viewsets.ModelViewSet):
@@ -66,7 +65,7 @@ class ContentSummaryViewSet(viewsets.ModelViewSet):
     serializer_class = ContentSummarySerializer
 
     def get_queryset(self):
-        return ContentNode.objects.all()
+        return ContentNode.objects.all().order_by('lft')
 
 
 class UserSummaryViewSet(viewsets.ModelViewSet):
@@ -96,13 +95,13 @@ class RecentReportViewSet(viewsets.ModelViewSet):
         self.kwargs['last_active_time'] = datetime_cutoff.isoformat()
         recent_content_items = ContentSummaryLog.objects.filter_by_topic(query_node).filter(
             Q(progress__gt=0) | Q(masterylogs__in=attempted_mastery_logs),
-            user__in=get_members_or_user(self.kwargs['collection_kind'], self.kwargs['collection_id']),
+            user__in=list(get_members_or_user(self.kwargs['collection_kind'], self.kwargs['collection_id'])),
             end_timestamp__gte=datetime_cutoff).values_list('content_id')
         # note from rtibbles:
         # As good as either I or jamalex could come up with to ensure that we only return
         # unique content_id'ed ContentNodes from the coach recent report endpoint.
         # Would have loved to use distinct('content_id'), but unfortunately DISTINCT ON is Postgresql only
         pks_with_unique_content_ids = ContentNode.objects.filter(
-            content_id__in=recent_content_items).values('content_id').annotate(
+            content_id__in=recent_content_items).values('content_id').order_by('lft').annotate(
             pk=Min('pk')).values_list('pk', flat=True)
-        return ContentNode.objects.filter(pk__in=pks_with_unique_content_ids)
+        return ContentNode.objects.filter(pk__in=pks_with_unique_content_ids).order_by('lft')