--- a/kolibri/plugins/coach/test/test_coach_api.py
+++ b/kolibri/plugins/coach/test/test_coach_api.py
@@ -88,8 +88,8 @@ class ContentReportAPITestCase(APITestCase):
             [{'log_count_total': 0, 'total_progress': 0.0, 'log_count_complete': 0}],
             [
                 {'kind': 'audio', 'node_count': 1, 'total_progress': 0.5},
-                {'kind': 'exercise', 'node_count': 1, 'total_progress': 0.7},
-                {'kind': 'document', 'node_count': 1, 'total_progress': 0}
+                {'kind': 'document', 'node_count': 1, 'total_progress': 0},
+                {'kind': 'exercise', 'node_count': 1, 'total_progress': 0.7}
             ]
         ])
         assert_progress(c2, [