## Summary

Optimizes database queries used in ContentReportSerializer to prevent excess queries and do more efficient queries.

Takes total queries for the tests I did down from 297 to 56, reduced the query time from ~360ms to ~220ms, and the CPU time for the request from ~4000ms to ~2200ms.

## TODO

- [X] Have tests been written for the new code?

## Issues addressed

Fixes #1658